handle compound arguments
Using getopts is the standard way of doing so. This handles flawlessly
even mixing options directly handled by makepkg-expanded with such
passed on to makepkg (for example -fg PKGBUILD).

Unfortunately, the routine gets quite complex as there are some caveats
especially with long arguments. Caution is mandatory when makepkg-
expanded should handle long options with arguments itself.

As getopts does not know about long arguments, they are processed as a
short option - with a mandatory argument (the rest of the -option), with
some error checking to reject weird constructions and typos like
-an-opt, which getopts would view as a compound of the short options -a
-n -- -o -p and -t, without even regarding -- as the stop signal in this
case.