deduplicate templates
Including a single template twice may have unforeseen consequences or at
least introduces redundancies into the expanded PKGBUILD. In an effort
to keep these neat and clean, a preprocessor now removes duplicate
templates from the build script entirely.

Note that this especially makes diamond inclusion convenient, where a
template A includes B and C which in turn both depend on D. To achieve
exactly one inclusion of D in A before, D had to be included in A
directly. The dependency of B and C on D had to be specified externally.
Alternatively, D had to specify countermeasures to executing it twice.
Now, the intermediates B and C can now simply specify the inclusion of D
to have it always present without worrying about its duplication.

The task could indeed be achieved more convenient by something like a
python script, but bash was chosen nonetheless to keep the script from
introducing a new dependency.