externalize version propagation
This is a first step to turn makepkg-expanded into a general processor
of expanded PKGBUILDs. The tasks it performs, such as building, pkgver
propagation and AUR preparation, could be freely chosen by specifying
their invocation as a future argument. To make that convenient, they
should all be compact command lines to specify. Already short are
makepkg and makesrc, now cp-pkgver joins them to complete the set of
tasks performed up to now.

It also may be useful on its own.