also copy pkgrel
When the tool was still part of makepkg-expanded, the pkgrel could
safely be assumed to be 1 whenever the pkgver changed, as it was only
run after a makepkg, which resets it to 1 whenever a pkgver function
yields a changed number. Now externalized, the routine should be able to
deal with more general cases where this is not the case, for example to
copy the version numbers from one package to another.