name branch according to pkgbase
This allows for multiple PKGBUILDs to be distributed from a single
repository, keeping them neatly assorted with branches below the aur
reference.

The pkgbase clearly identifies a package. It may change over time, for
example when the package is renamed or merged with others, but according
to AUR naming convention, packages will always be distributed to a
remote with the exact name of the current pkgbase.

The pkgbase variable is not required in each PKGBUILD. It can also be
substituted by the first pkgname. Thus, fetching this information from
the .SRCINFO, where this is already figured out, is much easier.