avoid hashing overwritten PKGBUILD
When an alternate build script is specified using the -p option, adding
a PKGBUILD file before introduces an object for this file to the
repository, which will never be relevant as it will be overwritten with
the alternate variant before composing the tree. Not hashing it
altogether not only saves work, but keeps the repository clean.