show changes when describing commit
Just like "git commit -v" does it, a difference status is shown below
the commit message to compose. As this commit is composed completely
automatic, this not only lets the user verify it, but also remind itself
of the changes about to port, so that they can be checked against the
proposed commit message.