reject empty messages
This not only gets rid of a hangup of git commit-tree, which seems to
wait on a commit message passed via the standard input when passed an
empty file through the -F option, but also gives the user a chance to
abort a commit. This is especially useful as the message by now shows
what it is about to be committed as a comment to verify.