utilize diff return code
Git's diff-index returns with a dedicated error code for wrong arguments
if a nonexistent revision is given. In the relevant case, this is the
branch name. So evaluating the exit code saves the call to git show-ref.
The error message issued by diff-index is suppressed as it is taken as
an indicator for further measure.

Another execution of diff-index can be saved by utilizing the invocation
of this command that is performed anyway when composing the commit
message template. This leaves only one case, namely the new creation of
the distribution branch, to invoke diff-index twice, which now also
completely rewrites the commit message. As this happens only once per
branch, it can be regarded as a corner case legitimating this
duplication in favor of saving an action in the more regular situation
of updating a branch.