reuse last commit message also on branch creation
Practical appliance showed that for majority of branch creation, the
last commit affecting the underlying files would be something like
"adopted package name" or "created package name", which serves as a
proper description for the first commit on the distribution branch also.
Thus, proposing this last commit not only when updating the branch but
also on creation not only simplifies the code by avoiding writing the
whole commit message draft twice, but also lends itself to more
expressive creation messages.