choose branch name
This is especially useful if there already is a distribution branch not
named as expected. It is also generally polite to let the user choose
the name for things he personally has to work with.

The subshell also has to be changed to a mere grouping so that the
assignment of the branch name variable can propagate out of it.