allow running multiple snippets
This makes it easier to specify a call sequence where every failure
shall propagate. In one consecutive script, error code propagation has
to be explicitly requested with something like "|| return $?", which can
be saved if this is the last call in a bash snippet, which will
determine the exit code of the bash invocation.