exclude original and generated file from globbing
In the majority of cases, processes in the code snippets passed to the
-r option are not intended to work on both the original PKGBUILD and the
expanded variant. Excluding them by default leads to more compact
snippets as they do not have to be manually excluded from wildcards,
which normally requires some bash trickery like the extglob shell option
or a call to the find program anyways.

In the seldom cases where both are to be considered, they can still be
specified explicitly.