extract AUR specific functionality
The majority of aurbranch code dealt with Git. Exporting all of this to
the dedicated offbranch program makes it into a general purpose tool to
commit files onto branches not checked out.

Functionality specific to the AUR is still retained under the original
name, which is now merely a wrapper around offbranch.