store preparation code in temporary file
Piping it saved writing this file, but blocked the standard input for
the process snippet. Storing to a temporary file instead frees it for
interactive tools withing the snippet to use.

It also has the benefit to write the preparation code only once per
PKGBUILD when multiple snippets are given, which is even the default.