resolve symbolic links
PKGBUILD collection repositories often keep reused files in a central
place and place links to them in the package directories. These links
would point to nowhere when submitted to the AUR as is. Including the
link target in their place is most probably preferred instead.

In case this is not the wanted behaviour, it can be circumvented still
by calling offbranch directly.