do not run outside repository
When there is no repository to commit to, running offbranch is useless
and will only lead to error messages.

Trickery with GIR_DIR and similar variables is still possible and
intended for advanced use. Tinkering with these should fairly save as
every modification to the repository is made through commands of the git
suite, which all honor these.