return verbose exit codes without -q
Suppressing offbranch's messages through aurbranch can be quite
confusing, especially when the user expects the commit to be aborted
because of an empty message, receives no confirmation and wonders
whether a commit actually happened.

As some commands exit with code 1 even when passed a -h flag to show
help, returning codes different from 0 to indicate partial success
should be fine.

This also slightly rewords the status messages about branch creation and
update to present tense because this can save a conditional block and a
variable.