tell apart templates
With another color, it is easier to visually capture the type of a node.

By placing the style directives individually after each line relation,
iterating the list of files twice would be saved. As this would result
in a quite cluttered output, one pass is invested to tell the categories
apart beforehand.