deal with package collections
When multiple packages are to be distributed via offremotes, they have
to have different names. The aurbranch program uses the location withing
the repository it is run from to determine whether it deals with such a
collection. Contrary to that, aurremote configures values global to the
repository, not tightly bound to any location, so this metric is
counterintuitive there. A switch is more fitting to make the choice.