leave pulling to wrapper
As aurbranch uses offremote directly, aurremote does not have to
distinguish between creating a new branch and adopting an existing
history. As it is safe to assume that aurremote is now only used to
adopt packages, this is a stronger implication for the intention to pull
than branch existence introduced by a former aurbranch execution.

To indicate this even more, the tool is also renamed to reflect adoption
of packages in its name.