check for conflicting branches
When dealing with hierarchical branch names that contain slashes, they
are mapped to directories and files below .git/refs. This makes it
impossible to have a branch aur when there already is a branch starting
with aur/ and vice versa.

As offremote will work fine in this case, the fetch command is the first
to fail. To not break with only the last step not correctly performed,
leaving the repository invoked on in an unexpected state, this is now
checked beforehand to stop any action to begin with.

It also implicitly warns the user when a -c flag is most probably
missing or superfluous.