specify message per option
This is especially useful when many commits shall be batch processed
with something like makepkg-expanded.

An option to avoid bringing up the editor is discouraged because to
cleanly implement it, the program should refrain from composing any
commit message to begin with, avoiding the need to later dissect it. As
the composition process is intertwined with checking on changed files
indicating commit necessity, this would make the code much harder to
read. Therefore, the possibility to still keep the editor from appearing
by setting it to something useless is only suggested in the manual page.