do not prioritize PKGBUILD last
Argument parsing in offbranch does not allow positional arguments
intermixed with options, because the file arguments can be rather
complicated and are thus better clearly separated. Because of this, to
pass arbitrary arguments to offbranch the variable collecting them had
to be moved before the PKGBUILD. When multiple files are specified to
offbranch with the same name, the last takes precedence in respect to
composing the revision. The result is that the user cannot overwrite the
PKGBUILD chosen through the -p option.

This was not yet documented and could be desired to just automatically
generate the .SRCINFO. Splitting options and files to pass enables this
now also documented feature again.