parallelize PKGBUILD and .SRCINFO option status
Searching the positional arguments for a file included with the name
.SRCINFO is a feat bash has to go quite great lengths to accomplish. It
is also unexpected due to the fact that the PKGBUILD is already handled
specially with an option.

Choosing both through options unifies the behaviour, leading to a clean
user expectation that all file arguments are passed through without
altering the behaviour of aurbranch. It also gives the chance to use a
.SRCINFO file only to name the branch and overwrite it for storage.