accept zero delimited file list
This makes it easier to pipe filenames with spaces and newlines.

An associative array maps filenames to the command suited to read them,
easily unifying arguments, files read with -i and streams read by -z all
into a single zero delimited list to be read by the usual read loop.