save repository root directory
Requesting Git to compute it multiple times is unnecessary. It will not
change, so it can be stored in a variable. This means that it would also
be computed when no arguments were given, but this generates a branch
without any files, which should not happen often enough to consider it
anyhow important.