guard -b shorthand against nullglob
When nothing matches the * wild card, the default settings of bash leave
the symbol unaltered, causing git-ls-files to indeed interpret it,
disregarding the GLOBIGNORE, normally finding the PKGBUILD again, using
it to overwrite the expanded one given to the -p switch. A combination
of setting the nullglob shopt, making an unmatched * disappear, with a
bait file will never be shown, namely the .git directory, keeps ls-files
expectedly silent in this case. This makes -b also applicable for
package directories that store nothing but the PKGBUILD in the package
directory.