suppress updates only affecting time stamp
Since the recent release of version 8, the mksrcinfo program adds a
header to each .SRCINFO including the date it was run. This introduces a
guaranteed change to the file, updating the branch unconditionally
whenever aurbranch is run. These unneeded commits should be avoided.

As all project using aurbranch are affected, it makes sense to implement
this here. However, be aware that this task can be handled by Git alone.
Using gitattributes(5), a clean filter can be defined affecting any
.SRCINFO file added to or updated in the repository to apply a sed
script fixing the date, for example to the last modification or commit
date of the PKGBUILD, or simply delete the second line containing the
time stamp.