find name from expanded PKGBUILD
The graph-templates program is only useful for PKGBUILDs that utilize
templates. If mksrcinfo is at all able to comprehend the build script
containing unexpanded templates it may still generate faulty
information. As templates can contain arbitrary code that may alter all
the variables, the only way to guarantee extraction of a valid pkgbase
is expanding the templates beforehand, which unfortunately comes with
the cost of a significant increase in processing time.