optimize directory extraction from empty path
The variable $pkgbuild, set by the -p option, has to be checked for
content whenever used. Expanding the default value substitution to a
complete if block can skip the call to dirname and the examination of
its result when all outcomes are predetermined through the default
value.

Assigning $pkgbuild and $srcinfo to their default when unset on first
use is carefully avoided not to pass an ineffective rename to offbranch.