use makepkg --printsrcinfo directly
In pkgbuild-introspecion commit a07cd59 [1], which was released with
version 8, mksrcinfo became a wrapper for makepkg --printsrcinfo. Using
it directly saves the dependency on the wrapper.

This also rids the automatically generated .SRCINFO files from the time
stamped comment, which was prone to unnecessarily dirty repositories and
thus circumvented in a dedicated commit.

[1]: https://github.com/falconindy/pkgbuild-introspection/commit/a07cd59cea847d182b500b0b17c7b7c750b3101c