style nodes individually
This increases the size of the output graph script, because instead of
one style definition for all templates and one for all build scripts,
the same definition is replicated on each node. But, carrying the style
with them, the nodes can now placed anywhere in the document, allowing
to process files in any order instead of templates and build scripts
separated.