integrate file processing into categorization
The template graphing routine is slimmed by processing all files
immediately as they are encountered, completely removing the need for
any arrays to collect them.

As argument processing is now completely interleaved with processing,
the graph header has to be already output before any argument is
examined. Thus, -h does not stop processing any longer but just
suppresses processing the defaults. The help text is also emitted on the
error stream to not interrupt an incomplete graph of nodes given beside
the flag.