parse .SRCINFO for additional files
Including all files placed in the same directory as the PKGBUILD as long
as they are tracked by Git is often the right approach for package
collections. In the second common use case, a project repository that
additionally stores a PKGBUILD, this PKGBUILD is normally the only file
relevant for the package, along with the generated .SRCINFO and possibly
few other files in comparison to the number of files beside it. When
makepkg-expanded blindly includes files when called with the -b switch,
project files are prone to slip into the package, needlessly duplicated.

Telling project packages apart from collections could be done by
examining the depth of the PKGBUILDs in the repository: Collections are
normally composed of directories collecting everything belonging to one
package each, while the PKGBUILD for a project often resides in the root
directory. The command string generated by the -b option of
makepkg-expanded could adapt according to this heuristic and thus
prevent accidental includes.

Moving the responsibility to determine which additional files to
distribute one step down in the stack of programs, aurbranch is in a
much better place to decide than makepkg-expanded, as it can extract
more knowledge about the package from the .SRCINFO. The source and
install elements tells much more clearly which files are meant to be
distributed with the PKGBUILD than guesswork beforehand.  This approach
promises to yield much more accurate predictions and reduce the number
of cases where manual adjustment is needed, which can still be done by
explicitly expressing the aurbranch call via the -r option to
makepkg-template.

For backwards compatibility, aurbranch acquired a new switch to restore
its old behaviour. Such a switch is not necessary in makepkg-expanded,
as the same can be achieved by passing the string composed by -b before
this commit to the -r option.